# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ShortUrl {
  id: number;
  userId: string;
  originalUrl: string;
  shortCode: string;
  /** @nullable */
  customAlias: string | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  expiresAt: string | null;
  isActive: boolean;
  totalClicks: number;
  /** @nullable */
  lastVisitedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UrlInput {
  originalUrl: string;
  customAlias?: string;
  title?: string;
  expiresAt?: string;
}

export interface UrlUpdate {
  originalUrl?: string;
  title?: string;
  customAlias?: string;
  expiresAt?: string;
  isActive?: boolean;
}

export interface UrlListResponse {
  urls: ShortUrl[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  totalUrls: number;
  totalClicks: number;
  activeUrls: number;
  clicksThisWeek: number;
  topUrl?: ShortUrl;
}

export interface ClickEvent {
  id: number;
  urlId: number;
  /** @nullable */
  deviceType: string | null;
  /** @nullable */
  browser: string | null;
  /** @nullable */
  os: string | null;
  /** @nullable */
  country: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  referrer: string | null;
  createdAt: string;
}

export interface AnalyticsSummary {
  urlId: number;
  shortCode: string;
  originalUrl: string;
  totalClicks: number;
  uniqueClicks: number;
  /** @nullable */
  lastVisitedAt: string | null;
  recentClicks: ClickEvent[];
}

export interface DailyClickPoint {
  date: string;
  clicks: number;
}

export interface BreakdownItem {
  label: string;
  count: number;
  percentage: number;
}

export interface AnalyticsBreakdown {
  devices: BreakdownItem[];
  browsers: BreakdownItem[];
  os: BreakdownItem[];
  countries: BreakdownItem[];
}

export interface RedirectResponse {
  originalUrl: string;
  /** @nullable */
  title: string | null;
}

export interface PublicStats {
  shortCode: string;
  /** @nullable */
  title: string | null;
  totalClicks: number;
  createdAt: string;
  dailyClicks: DailyClickPoint[];
}

export type ListUrlsParams = {
page?: number;
limit?: number;
search?: string;
};