# lib/api-zod/src/generated/types/analyticsBreakdown.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { BreakdownItem } from './breakdownItem';

export interface AnalyticsBreakdown {
  devices: BreakdownItem[];
  browsers: BreakdownItem[];
  os: BreakdownItem[];
  countries: BreakdownItem[];
}