# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { ClickEvent } from './clickEvent';

export interface AnalyticsSummary {
  urlId: number;
  shortCode: string;
  originalUrl: string;
  totalClicks: number;
  uniqueClicks: number;
  /** @nullable */
  lastVisitedAt: string | null;
  recentClicks: ClickEvent[];
}