# lib/api-zod/src/generated/types/clickEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface ClickEvent {
  id: number;
  urlId: number;
  /** @nullable */
  deviceType: string | null;
  /** @nullable */
  browser: string | null;
  /** @nullable */
  os: string | null;
  /** @nullable */
  country: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  referrer: string | null;
  createdAt: string;
}