# lib/api-zod/src/generated/types/dailyClickPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyClickPoint {
  date: string;
  clicks: number;
}