# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { ShortUrl } from './shortUrl';

export interface DashboardStats {
  totalUrls: number;
  totalClicks: number;
  activeUrls: number;
  clicksThisWeek: number;
  topUrl?: ShortUrl;
}