# lib/api-zod/src/generated/types/publicStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyClickPoint } from './dailyClickPoint';

export interface PublicStats {
  shortCode: string;
  /** @nullable */
  title: string | null;
  totalClicks: number;
  createdAt: string;
  dailyClicks: DailyClickPoint[];
}