# lib/api-zod/src/generated/types/shortUrl.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface ShortUrl {
  id: number;
  userId: string;
  originalUrl: string;
  shortCode: string;
  /** @nullable */
  customAlias: string | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  expiresAt: string | null;
  isActive: boolean;
  totalClicks: number;
  /** @nullable */
  lastVisitedAt: string | null;
  createdAt: string;
  updatedAt: string;
}