# lib/api-zod/src/generated/types/urlInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface UrlInput {
  originalUrl: string;
  customAlias?: string;
  title?: string;
  expiresAt?: string;
}