# lib/api-zod/src/generated/types/urlListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { ShortUrl } from './shortUrl';

export interface UrlListResponse {
  urls: ShortUrl[];
  total: number;
  page: number;
  limit: number;
}