# lib/api-zod/src/generated/types/urlUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Shortener with Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface UrlUpdate {
  originalUrl?: string;
  title?: string;
  customAlias?: string;
  expiresAt?: string;
  isActive?: boolean;
}